Remove Policy/Enforcer split
